Rename CompactCalendar to MiniCalendar, fix template and sticky state bugs
  Refactor
  - Rename CompactCalendar to MiniCalendar throughout codebase
  - Update all references in constants, settings, templates, and localization
  - Delete old compact-calendar.mjs, compact-calendar.hbs, tab-compact.hbs
  - Add new mini-calendar.mjs, mini-calendar.hbs, tab-mini-calendar.hbs

  Bug Fixes
  - Fix duplicate nested structure in tab-months.hbs causing month names to appear twice
  - Fix MiniCalendar sticky states not applying when changed in settings panel
  - Add MiniCalendar.refreshStickyStates() to update instance when settings change
  - Update #restoreStickyStates() to properly toggle classes on/off

  Localization
  - Rename all CALENDARIA.CompactCalendar.* keys to CALENDARIA.MiniCalendar.*
  - Add missing CALENDARIA.Repeat.computed localization string

  Documentation
  - Update API-Reference, Calendar-HUD, Getting-Started, Settings, Troubleshooting
  - Add GitHub assets folder
