+++ lang/pt-BR.json
@@ -849,15 +849,15 @@
   "CALENDARIA.Common.Weekday": "Dia de Semana",
   "CALENDARIA.Common.Weekdays": "Dias de Semana",
   "CALENDARIA.Common.Year": "Ano",
-  "CALENDARIA.CompactCalendar.AllDay": "O Dia Todo",
-  "CALENDARIA.CompactCalendar.CloseNotesPanel": "Fechar Notas",
-  "CALENDARIA.CompactCalendar.OpenFull": "Abrir Calendário Completo",
-  "CALENDARIA.CompactCalendar.PinControls": "Fixar Controles",
-  "CALENDARIA.CompactCalendar.SetCurrentDate": "Definir Data Atual",
-  "CALENDARIA.CompactCalendar.StickyPosition": "Travar: Posição Bloqueada",
-  "CALENDARIA.CompactCalendar.StickySidebar": "Travar: Barra Lateral",
-  "CALENDARIA.CompactCalendar.StickyTimeControls": "Travar: Controles de Tempo",
-  "CALENDARIA.CompactCalendar.ViewNotes": "Ver Notas",
+  "CALENDARIA.MiniCalendar.AllDay": "O Dia Todo",
+  "CALENDARIA.MiniCalendar.CloseNotesPanel": "Fechar Notas",
+  "CALENDARIA.MiniCalendar.OpenFull": "Abrir Calendário Completo",
+  "CALENDARIA.MiniCalendar.PinControls": "Fixar Controles",
+  "CALENDARIA.MiniCalendar.SetCurrentDate": "Definir Data Atual",
+  "CALENDARIA.MiniCalendar.StickyPosition": "Travar: Posição Bloqueada",
+  "CALENDARIA.MiniCalendar.StickySidebar": "Travar: Barra Lateral",
+  "CALENDARIA.MiniCalendar.StickyTimeControls": "Travar: Controles de Tempo",
+  "CALENDARIA.MiniCalendar.ViewNotes": "Ver Notas",
   "CALENDARIA.ContextMenu.DeleteConfirm": "Excluir nota \"{name}\"?",
   "CALENDARIA.ContextMenu.DeleteNote": "Excluir Nota",
   "CALENDARIA.ContextMenu.EditNote": "Editar Nota",
@@ -1650,8 +1650,8 @@
   "CALENDARIA.Settings.ChatTimestampMode.Replace": "Substituir em tempo real",
   "CALENDARIA.Settings.ChatTimestampShowTime.Hint": "Inclua o tempo do jogo (horas:minutos) junto com a data nos registros de tempo do chat.",
   "CALENDARIA.Settings.ChatTimestampShowTime.Name": "Exibir Horário nos Registros de Tempo",
-  "CALENDARIA.Settings.CompactControlsDelay.Hint": "Quanto tempo (em segundos) leva para a barra lateral e os controles de tempo do calendário compacto se ocultarem automaticamente após o cursor do mouse sair do local.",
-  "CALENDARIA.Settings.CompactControlsDelay.Name": "Controles de Ocultação Automática e Atraso",
+  "CALENDARIA.Settings.MiniCalendarControlsDelay.Hint": "Quanto tempo (em segundos) leva para a barra lateral e os controles de tempo do MiniCalendar se ocultarem automaticamente após o cursor do mouse sair do local.",
+  "CALENDARIA.Settings.MiniCalendarControlsDelay.Name": "Controles de Ocultação Automática e Atraso",
   "CALENDARIA.Settings.DarknessSync.Hint": "Sincroniza automaticamente os níveis de escuridão da cena com a posição do sol com base na hora atual. Esta é a configuração padrão para novas cenas.",
   "CALENDARIA.Settings.DarknessSync.Name": "Sincronizar Escuridão da Cena com Tempo",
   "CALENDARIA.Settings.Importer.Hint": "Importe calendários do Simple Calendar, fantasy-calendar.com e outras fontes.",
@@ -1669,8 +1669,8 @@
   "CALENDARIA.Settings.ResetPosition.Success": "A posição do calendário foi redefinida.",
   "CALENDARIA.Settings.ShowCalendarHUD.Hint": "Exibir a HUD do Calendário ao carregar o mundo.",
   "CALENDARIA.Settings.ShowCalendarHUD.Name": "Exibir HUD do Calendário ao Carregar",
-  "CALENDARIA.Settings.ShowCompactCalendar.Hint": "Exibe o Calendário Compacto ao carregar o mundo.",
-  "CALENDARIA.Settings.ShowCompactCalendar.Name": "Exibir Calendário Compacto ao Carregar",
+  "CALENDARIA.Settings.ShowMiniCalendar.Hint": "Exibe o MiniCalendar ao carregar o mundo.",
+  "CALENDARIA.Settings.ShowMiniCalendar.Name": "Exibir MiniCalendar ao Carregar",
   "CALENDARIA.Settings.ShowMoonPhases.Hint": "Exibir indicadores de fases da lua no cabeçalho do calendário, mostrando a fase atual de cada lua configurada.",
   "CALENDARIA.Settings.ShowMoonPhases.Name": "Mostrar Fases da Lua",
   "CALENDARIA.Settings.ShowTimeKeeper.Hint": "Exibe a HUD do TimeKeeper ao carregar o mundo, para acesso rápido ao controle do tempo.",
@@ -1710,7 +1710,7 @@
   "CALENDARIA.SettingsPanel.Hint.LockHUDPosition": "Impedir que a HUD do Calendário seja arrastado para uma nova posição.",
   "CALENDARIA.SettingsPanel.Hint.ResetPositions": "Redefina todos os elementos da interface do Calendaria para suas posições padrão na tela.",
   "CALENDARIA.SettingsPanel.Hint.StickyHUDPosition": "Lembrar da posição do HUD entre as sessões.",
-  "CALENDARIA.SettingsPanel.Hint.StickyPosition": "Impedir que o calendário compacto seja arrastado para uma nova posição.",
+  "CALENDARIA.SettingsPanel.Hint.StickyPosition": "Impedir que o MiniCalendar seja arrastado para uma nova posição.",
   "CALENDARIA.SettingsPanel.Hint.StickySidebar": "Mantenha a barra lateral visível em vez de ocultá-la automaticamente quando o cursor sair do lugar.",
   "CALENDARIA.SettingsPanel.Hint.StickyStates": "Ative essas opções para manter os elementos da interface visíveis em vez de ocultá-los automaticamente.",
   "CALENDARIA.SettingsPanel.Hint.StickyTimeControls": "Mantenha os controles de tempo visíveis em vez de ocultá-los automaticamente quando o cursor do mouse sair do lugar.",
@@ -1729,7 +1729,7 @@
   "CALENDARIA.SettingsPanel.Section.Categories": "Categorias de Notas",
   "CALENDARIA.SettingsPanel.Section.ChatTimestamps": "Registro de Tempo do Chat",
   "CALENDARIA.SettingsPanel.Section.Climate": "Configuração Climática",
-  "CALENDARIA.SettingsPanel.Section.CompactCalendar": "Calendário Compacto",
+  "CALENDARIA.SettingsPanel.Section.MiniCalendar": "MiniCalendar",
   "CALENDARIA.SettingsPanel.Section.Debugging": "Debugging",
   "CALENDARIA.SettingsPanel.Section.Display": "Opções de Exibição",
   "CALENDARIA.SettingsPanel.Section.Position": "Posição",
@@ -1747,7 +1747,7 @@
   "CALENDARIA.SettingsPanel.Tab.Advanced": "Avançado",
   "CALENDARIA.SettingsPanel.Tab.Appearance": "Aparência",
   "CALENDARIA.SettingsPanel.Tab.Chat": "Chat",
-  "CALENDARIA.SettingsPanel.Tab.Compact": "Calendário Compacto",
+  "CALENDARIA.SettingsPanel.Tab.MiniCalendar": "MiniCal",
   "CALENDARIA.SettingsPanel.Tab.FullCal": "Calendário Completo",
   "CALENDARIA.SettingsPanel.Tab.HUD": "HUD",
   "CALENDARIA.SettingsPanel.Tab.Macros": "Macros",
