+++ scripts/calendar/calendar-manager.mjs
@@ -234,7 +234,7 @@ export default class CalendarManager {
   static #rerenderCalendarUIs() {
     for (const app of foundry.applications.instances.values()) {
       const name = app.constructor.name;
-      if (['CalendariaHUD', 'TimeKeeperHUD', 'CompactCalendar', 'CalendarApplication'].includes(name)) app.render();
+      if (['CalendariaHUD', 'TimeKeeperHUD', 'MiniCalendar', 'CalendarApplication'].includes(name)) app.render();
     }
   }
