+++ scripts/constants.mjs
@@ -29,12 +29,12 @@ export const SETTINGS = {
   THEME_MODE: 'themeMode',
   SHOW_TIME_KEEPER: 'showTimeKeeper',
   TIME_KEEPER_POSITION: 'timeKeeperPosition',
-  SHOW_COMPACT_CALENDAR: 'showCompactCalendar',
+  SHOW_MINI_CALENDAR: 'showMiniCalendar',
   ADVANCE_TIME_ON_COMBAT: 'advanceTimeOnCombat',
   SYNC_CLOCK_PAUSE: 'syncClockPause',
-  COMPACT_CALENDAR_POSITION: 'compactCalendarPosition',
-  COMPACT_CONTROLS_DELAY: 'compactControlsDelay',
-  COMPACT_STICKY_STATES: 'compactStickyStates',
+  MINI_CALENDAR_POSITION: 'miniCalendarPosition',
+  MINI_CALENDAR_CONTROLS_DELAY: 'miniCalendarControlsDelay',
+  MINI_CALENDAR_STICKY_STATES: 'miniCalendarStickyStates',
   DEV_MODE: 'devMode',
   CURRENT_WEATHER: 'currentWeather',
   CUSTOM_WEATHER_PRESETS: 'customWeatherPresets',
@@ -58,8 +58,8 @@ export const SETTINGS = {
 export const DISPLAY_LOCATIONS = {
   HUD_DATE: 'hudDate',
   HUD_TIME: 'hudTime',
-  COMPACT_HEADER: 'compactHeader',
-  COMPACT_TIME: 'compactTime',
+  MINI_CALENDAR_HEADER: 'miniCalendarHeader',
+  MINI_CALENDAR_TIME: 'miniCalendarTime',
   FULL_CALENDAR_HEADER: 'fullCalendarHeader',
   CHAT_TIMESTAMP: 'chatTimestamp'
 };
@@ -82,7 +82,7 @@ export const TEMPLATES = {
     PANEL_MACROS: `modules/${MODULE.ID}/templates/settings/tab-macros.hbs`,
     PANEL_CHAT: `modules/${MODULE.ID}/templates/settings/tab-chat.hbs`,
     PANEL_ADVANCED: `modules/${MODULE.ID}/templates/settings/tab-advanced.hbs`,
-    PANEL_COMPACT: `modules/${MODULE.ID}/templates/settings/tab-compact.hbs`,
+    PANEL_MINI_CALENDAR: `modules/${MODULE.ID}/templates/settings/tab-mini-calendar.hbs`,
     PANEL_HUD: `modules/${MODULE.ID}/templates/settings/tab-hud.hbs`,
     PANEL_TIMEKEEPER: `modules/${MODULE.ID}/templates/settings/tab-timekeeper.hbs`,
     PANEL_FULLCAL: `modules/${MODULE.ID}/templates/settings/tab-fullcal.hbs`,
@@ -95,7 +95,7 @@ export const TEMPLATES = {
   },
   TIME_DIAL: `modules/${MODULE.ID}/templates/time-dial.hbs`,
   TIME_KEEPER_HUD: `modules/${MODULE.ID}/templates/time-keeper-hud.hbs`,
-  COMPACT_CALENDAR: `modules/${MODULE.ID}/templates/compact-calendar.hbs`,
+  MINI_CALENDAR: `modules/${MODULE.ID}/templates/mini-calendar.hbs`,
   CALENDAR_HUD: `modules/${MODULE.ID}/templates/calendaria-hud.hbs`,
   CALENDAR_HUD_DOME: `modules/${MODULE.ID}/templates/calendaria-hud-dome.hbs`,
   CALENDAR_HUD_BAR: `modules/${MODULE.ID}/templates/calendaria-hud-bar.hbs`,
