+++ scripts/utils/theme-utils.mjs
@@ -133,7 +133,7 @@ export const COMPONENT_CATEGORIES = {
   common: 'CALENDARIA.ThemeEditor.Component.Common',
   domeHud: 'CALENDARIA.ThemeEditor.Component.DomeHud',
   timeKeeper: 'CALENDARIA.ThemeEditor.Component.TimeKeeper',
-  compactCalendar: 'CALENDARIA.ThemeEditor.Component.CompactCalendar'
+  miniCalendar: 'CALENDARIA.ThemeEditor.Component.MiniCalendar'
 };
 
 /**
@@ -155,9 +155,9 @@ export const COLOR_DEFINITIONS = [
   { key: 'text', label: 'CALENDARIA.ThemeEditor.Colors.Text', category: 'text', component: 'common' },
   { key: 'textDim', label: 'CALENDARIA.ThemeEditor.Colors.TextDim', category: 'text', component: 'common' },
   { key: 'titleText', label: 'CALENDARIA.ThemeEditor.Colors.TitleText', category: 'text', component: 'common' },
-  { key: 'weekdayHeader', label: 'CALENDARIA.ThemeEditor.Colors.WeekdayHeader', category: 'text', component: 'compactCalendar' },
-  { key: 'dayNumber', label: 'CALENDARIA.ThemeEditor.Colors.DayNumber', category: 'text', component: 'compactCalendar' },
-  { key: 'restDay', label: 'CALENDARIA.ThemeEditor.Colors.RestDay', category: 'text', component: 'compactCalendar' },
+  { key: 'weekdayHeader', label: 'CALENDARIA.ThemeEditor.Colors.WeekdayHeader', category: 'text', component: 'miniCalendar' },
+  { key: 'dayNumber', label: 'CALENDARIA.ThemeEditor.Colors.DayNumber', category: 'text', component: 'miniCalendar' },
+  { key: 'restDay', label: 'CALENDARIA.ThemeEditor.Colors.RestDay', category: 'text', component: 'miniCalendar' },
 
   // Buttons
   { key: 'buttonBg', label: 'CALENDARIA.ThemeEditor.Colors.ButtonBackground', category: 'buttons', component: 'common' },
