--- templates/settings/tab-compact.hbs
@@ -1,72 +1,0 @@
-<section class="tab {{tab.cssClass}}" data-tab="{{tab.id}}" data-group="{{tab.group}}">
-
-  <div class="button-row open-app-button">
-    <button type="button" data-action="openCompact">
-      <i class="fas fa-external-link-alt"></i>
-      {{localize "CALENDARIA.Common.Open"}}
-    </button>
-  </div>
-
-  <fieldset>
-    <legend>{{localize "CALENDARIA.SettingsPanel.Section.Display"}}</legend>
-
-    <div class="form-group">
-      <label>{{localize "CALENDARIA.Settings.ShowCompactCalendar.Name"}}</label>
-      <div class="form-fields">
-        <input type="checkbox" name="showCompactCalendar" {{#if showCompactCalendar}}checked{{/if}}>
-      </div>
-      <p class="hint">{{localize "CALENDARIA.Settings.ShowCompactCalendar.Hint"}}</p>
-    </div>
-
-    <div class="form-group">
-      <label>{{localize "CALENDARIA.Settings.CompactControlsDelay.Name"}}</label>
-      <div class="form-fields">
-        <input type="range" name="compactControlsDelay" value="{{compactControlsDelay}}" min="1" max="10" step="1">
-        <span class="range-value">{{compactControlsDelay}}s</span>
-      </div>
-      <p class="hint">{{localize "CALENDARIA.Settings.CompactControlsDelay.Hint"}}</p>
-    </div>
-  </fieldset>
-
-  <fieldset>
-    <legend>{{localize "CALENDARIA.SettingsPanel.Section.StickyStates"}}</legend>
-    <input type="hidden" name="compactStickySection" value="1">
-    <p class="hint section-hint">{{localize "CALENDARIA.SettingsPanel.Hint.StickyStates"}}</p>
-
-    <div class="form-group">
-      <label>{{localize "CALENDARIA.Common.StickyTimeControls"}}</label>
-      <div class="form-fields">
-        <input type="checkbox" name="compactStickyTimeControls" {{#if compactStickyTimeControls}}checked{{/if}}>
-      </div>
-      <p class="hint">{{localize "CALENDARIA.SettingsPanel.Hint.StickyTimeControls"}}</p>
-    </div>
-
-    <div class="form-group">
-      <label>{{localize "CALENDARIA.SettingsPanel.Sticky.Sidebar"}}</label>
-      <div class="form-fields">
-        <input type="checkbox" name="compactStickySidebar" {{#if compactStickySidebar}}checked{{/if}}>
-      </div>
-      <p class="hint">{{localize "CALENDARIA.SettingsPanel.Hint.StickySidebar"}}</p>
-    </div>
-
-    <div class="form-group">
-      <label>{{localize "CALENDARIA.Common.LockPosition"}}</label>
-      <div class="form-fields">
-        <input type="checkbox" name="compactStickyPosition" {{#if compactStickyPosition}}checked{{/if}}>
-      </div>
-      <p class="hint">{{localize "CALENDARIA.SettingsPanel.Hint.StickyPosition"}}</p>
-    </div>
-  </fieldset>
-
-  <fieldset>
-    <legend>{{localize "CALENDARIA.SettingsPanel.Section.Position"}}</legend>
-
-    <div class="button-row">
-      <button type="button" data-action="resetPosition" data-target="compact">
-        <i class="fas fa-undo"></i>
-        {{localize "CALENDARIA.Common.ResetPosition"}}
-      </button>
-    </div>
-  </fieldset>
-
-</section>
