+++ templates/sheets/calendar-note-form.hbs
@@ -88,7 +88,7 @@
         <label class="all-day-label">
           <input type="checkbox" name="system.allDay" {{checked
             system.allDay}} data-tooltip="CALENDARIA.UI.AllDayTooltip">
-          {{localize "CALENDARIA.CompactCalendar.AllDay"}}
+          {{localize "CALENDARIA.MiniCalendar.AllDay"}}
         </label>
         <label>{{localize "CALENDARIA.Common.End"}}</label>
         <input type="time" name="system.endDate.time" value="{{endTimeValue}}" {{#if system.allDay}}disabled{{/if}}
